fix: 피그마에 디자인된대로 사이드바를 구현합니다. #10
