+++ .commitlintrc.json
@@ -1,0 +1,32 @@
+{
+  "extends": ["@commitlint/config-conventional"],
+  "rules": {
+    "type-case": [2, "always", "lower-case"],
+    "type-empty": [2, "never"],
+    "type-enum": [
+      2,
+      "always",
+      [
+        "feat",
+        "fix",
+        "hotfix",
+        "docs",
+        "style",
+        "refactor",
+        "perf",
+        "design",
+        "test",
+        "settings",
+        "build",
+        "ci",
+        "chore",
+        "comment",
+        "rename",
+        "remove",
+        "revert"
+      ]
+    ],
+    "subject-empty": [2, "never"],
+    "subject-full-stop": [2, "never", "."]
+  }
+}
