+++ src/components/Layout/Nav/index.tsx
@@ -1,0 +1,12 @@
+import { Outlet } from 'react-router-dom';
+
+function Nav() {
+  return (
+    <div>
+      <h1>Nav</h1>
+      <Outlet />
+    </div>
+  );
+}
+
+export default Nav;
