--- src/components/sidebar/index.ts
@@ -1,3 +1,0 @@
-import { default as SideBar } from "./SideBar"
-
-export { SideBar };
